requirements.txt
ultralytics==8.3.128
viam-sdk==0.47.0

torch==2.7.0
torchvision==0.22.0
typing-extensions
